requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-debug-toolbar==3.2.1
idna==3.2
Pillow==8.3.1
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
slugify==0.0.1
sqlparse==0.4.1
text-unidecode==1.3
Unidecode==1.2.0
urllib3==1.26.6
